feat : 자연어 입력 ai 처리 과정 인증 로직 추가
