+++ src/main/kotlin/com/Dnight/calinify/ai_process/service/EventProcessingService.kt
@@ -3,9 +3,8 @@ package com.dnight.calinify.ai_process.service
 import com.dnight.calinify.ai_process.dto.request.PlainTextProcessingRequestDTO
 import com.dnight.calinify.ai_process.dto.response.ProcessedEventResponseDTO
 import com.dnight.calinify.ai_process.dto.to_ai.request.AiPlainTextProcessingRequestDTO
-import com.dnight.calinify.ai_process.dto.to_ai.request.AiRequestDTO
 import com.dnight.calinify.ai_process.dto.to_ai.response.AiPlainTextProcessedResponseDTO
-import com.dnight.calinify.ai_process.dto.to_ai.response.AiResponseDTO
+import com.dnight.calinify.ai_process.module.PlainTextRequest
 import com.dnight.calinify.ai_process.repository.AiProcessingEventRepository
 import com.dnight.calinify.ai_process.repository.AiProcessingStatisticsRepository
 import com.dnight.calinify.config.basicResponse.ResponseCode
@@ -16,65 +15,37 @@ import com.dnight.calinify.user.repository.UserRepository
 import jakarta.transaction.Transactional
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
 import org.springframework.stereotype.Service
-import org.springframework.web.reactive.function.client.WebClient
-import org.springframework.web.reactive.function.client.WebClientRequestException
-import reactor.core.publisher.Mono
-import java.net.ConnectException
-import kotlin.reflect.KClass
 
 @Service
 class EventProcessingService(
     private val aiProcessingEventRepository: AiProcessingEventRepository,
     private val aiProcessingStatisticsRepository: AiProcessingStatisticsRepository,
     private val userRepository: UserRepository,
-    aiRequestBuilder: WebClient.Builder
+    private val plainTextRequest: PlainTextRequest
 ) {
-    private val aiRequestClient = aiRequestBuilder.build()
-
-    fun <T : AiResponseDTO> aiRequest(requestDTO : AiRequestDTO, responseCls : KClass<T>) : ResponseEntity<T>? {
-        return aiRequestClient.post()
-            .uri("/api/v1/plainText")
-            .bodyValue(requestDTO)
-            .retrieve()
-            .toEntity(responseCls.java)
-            .onErrorResume { ex ->
-                when (ex) {
-                    is WebClientRequestException -> {
-                        Mono.error(ServerSideException(ResponseCode.AiRequestFail))
-                    }
-                    is ConnectException -> {
-                        Mono.error(ServerSideException(ResponseCode.AiRequestFail))
-                    }
-                    else -> {
-                        Mono.error(RuntimeException("알 수 없는 오류 발생: ${ex.message}"))
-                    }
-                }
-            }
-            .block()
-    }
 
     @Transactional(dontRollbackOn = [DontRollbackException::class])
-    fun createPlainTextEvent(plainTextProcessingRequestDTO: PlainTextProcessingRequestDTO) : ProcessedEventResponseDTO{
+    fun createPlainTextEvent(plainTextProcessingRequestDTO: PlainTextProcessingRequestDTO,
+                             userId : Long) : ProcessedEventResponseDTO{
 
         // DTO로부터 ai server에 전달할 내용을 정제
         val aiPlainTextProcessingRequestDTO = AiPlainTextProcessingRequestDTO.toEntity(plainTextProcessingRequestDTO)
 
         // ai server request
-        val aiPlainTextProcessedResponseDTO = aiRequest(aiPlainTextProcessingRequestDTO, AiPlainTextProcessedResponseDTO::class)
+        val aiPlainTextProcessedResponseDTO = plainTextRequest.aiRequest(aiPlainTextProcessingRequestDTO, AiPlainTextProcessedResponseDTO::class)
             ?: throw ServerSideException(ResponseCode.AiRequestFail)
 
         // 응답으로부터 body 추출
         val processedResponseBody = aiPlainTextProcessedResponseDTO.body!!
 
-        // TODO User 정보는 더미 데이터, 토큰에서 유저 값 갖고 오기
-        val userEntity = userRepository.findByIdOrNull(1) ?: throw ClientException(ResponseCode.UserNotFound)
+        // User 조회
+        val userEntity = userRepository.findByIdOrNull(userId) ?: throw ClientException(ResponseCode.UserNotFound)
 
         // 통계 객체 생성
         val eventProcessedStatisticsEntity = AiPlainTextProcessedResponseDTO.toStatisticsEntity(userEntity, processedResponseBody, plainTextProcessingRequestDTO)
 
-        // 일정 정보 파악 실패 시 예외 발생
+        // 일정 정보 파악 실패 시 예외 발생, 실패 경우 저장을 위해 트랜잭션 롤백 X
         if (aiPlainTextProcessedResponseDTO.statusCode.isSameCodeAs(HttpStatus.ACCEPTED)) {
             eventProcessedStatisticsEntity.isSuccess = 0
             aiProcessingStatisticsRepository.save(eventProcessedStatisticsEntity)
@@ -89,7 +60,7 @@ class EventProcessingService(
         val processingEventEntity = AiPlainTextProcessedResponseDTO.toEventEntity(eventProcessingId, processedResponseBody)
         aiProcessingEventRepository.save(processingEventEntity)
 
-        // 프로세싱 결과에 statistics, event id를 삽입하여 반환
+        // 프로세싱 결과에 event id, statistics를 삽입하여 반환
         val processedEventResponseDTO = ProcessedEventResponseDTO.from(
             eventProcessingId, processedResponseBody)
